+++ src/hittable/sphere.rs
@@ -9,18 +9,18 @@ use super::Hittable;
 #[derive(Debug, Default)]
 pub struct Sphere {
     pub center: Point3,
-    pub radius: f32,
+    pub radius: f64,
     pub material: Material
 }
 
 impl Sphere {
-    pub fn new(center: Point3, radius: f32, material: Material) -> Self {
+    pub fn new(center: Point3, radius: f64, material: Material) -> Self {
         Self { center, radius, material}
     }
 }
 
 impl Hittable for Sphere {
-    fn hit(&self, r: Ray, t_min: f32, t_max: f32, rec: &mut HitRecord) -> bool {
+    fn hit(&self, r: Ray, t_min: f64, t_max: f64, rec: &mut HitRecord) -> bool {
         let oc = r.origin() - self.center;
         let a = r.dir().norm_squared();
         let half_b = dot(oc, r.dir());
@@ -30,10 +30,10 @@ impl Hittable for Sphere {
             return false;
         }
 
-        let mut root = (-half_b - f32::sqrt(discriminant)) / a;
+        let mut root = (-half_b - f64::sqrt(discriminant)) / a;
 
         if root < t_min || root > t_max {
-            root = (-half_b + f32::sqrt(discriminant)) / a;
+            root = (-half_b + f64::sqrt(discriminant)) / a;
             if root < t_min || root > t_max {
                 return false;
             }
