+++ src/ray/functions.rs
@@ -7,7 +7,7 @@ impl Ray {
         Self { origin, dir }
     }
 
-    pub fn at(&self, t: f32) -> Point3 {
+    pub fn at(&self, t: f64) -> Point3 {
         self.origin + t * self.dir
     }
